+++ src/main/java/home/mutant/dl/gaussian/pattern/MLRunGaussianMBPatternNet.java
@@ -62,8 +62,8 @@ public class MLRunGaussianMBPatternNet {
 	}
 
 	public static void replaceImage(List<MBPatternNeuron> neurons, int indexImage, List<Image> images){
-		float[] newImage = new float[neurons.size()];
-		float[] pixels = images.get(indexImage).data;
+		double[] newImage = new double[neurons.size()];
+		double[] pixels = images.get(indexImage).data;
 		for (int j = 0; j < neurons.size(); j++) {
 			int countSimilar = (int) neurons.get(j).countSimilar(pixels);
 			newImage[j] = (byte) (countSimilar>100?countSimilar:0);
@@ -199,7 +199,7 @@ public class MLRunGaussianMBPatternNet {
 		//MBPatternNeuron.similarityRate = 0.75;
 		for (int i=0;i<Globals.NO_TRAIN;i++)
 		{
-			float[] pixels = MnistDatabase.trainImages.get(i).data;
+			double[] pixels = MnistDatabase.trainImages.get(i).data;
 			for (int j = 0; j < allNeurons.size(); j++) {
 				int countSimilar = (int) allNeurons.get(j).countSimilar(pixels);
 				input[j] = countSimilar>=Globals.SUBIMG_SIZE*Globals.SIMILARITY?countSimilar:0;
@@ -218,7 +218,7 @@ public class MLRunGaussianMBPatternNet {
 		int count=0;
 		for (int i=0;i<Globals.NO_TEST;i++)
 		{
-			float[] pixels = MnistDatabase.testImages.get(i).data;
+			double[] pixels = MnistDatabase.testImages.get(i).data;
 			for (int j = 0; j < allNeurons.size(); j++) {
 				int countSimilar = (int) allNeurons.get(j).countSimilar(pixels);
 				input[j] = countSimilar>=Globals.SUBIMG_SIZE*Globals.SIMILARITY?countSimilar:0;
@@ -243,7 +243,7 @@ public class MLRunGaussianMBPatternNet {
 		//MBPatternNeuron.similarityRate = 0.75;
 		for (int i=0;i<Globals.NO_TRAIN;i++)
 		{
-			float[] pixels = MnistDatabase.trainImages.get(i).data;
+			double[] pixels = MnistDatabase.trainImages.get(i).data;
 			for (int j = 0; j < allNeurons.size(); j++) {
 				int countSimilar = (int) allNeurons.get(j).countSimilar(pixels);
 				input[j] = countSimilar>=Globals.SUBIMG_SIZE*Globals.SIMILARITY?countSimilar:0;
@@ -262,7 +262,7 @@ public class MLRunGaussianMBPatternNet {
 		int count=0;
 		for (int i=0;i<Globals.NO_TEST;i++)
 		{
-			float[] pixels = MnistDatabase.testImages.get(i).data;
+			double[] pixels = MnistDatabase.testImages.get(i).data;
 			for (int j = 0; j < allNeurons.size(); j++) {
 				int countSimilar = (int) allNeurons.get(j).countSimilar(pixels);
 				input[j] = countSimilar>=Globals.SUBIMG_SIZE*Globals.SIMILARITY?countSimilar:0;
